requirements.txt
streamlit
ffmpeg-python
openai-whisper
opencv-python
openai
nltk
matplotlib
pandas
numpy
moviepy
imageio-ffmpeg
